pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
beautifulsoup4 = "^4.11.1"
requests="^2.28.1"
validators = "^0.20.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
